# internal/server/oapi/swagger.yaml
components:
  schemas:
...
    server.createJobRequest:
      properties:
        distribute:
          description: Engine store names that should pull the image afterwards.
          example:
          - node-a
          - node-b
          items:
            type: string
          type: array
          uniqueItems: false
        from:
          description: Source registry store name or host; defaults to the ref's registry.
          example: dockerhub
          type: string
        platforms:
          description: Platforms to move; defaults to the server platform when empty.
          example:
          - linux/amd64
          - linux/arm64
          items:
            type: string
          type: array
          uniqueItems: false
        ref:
          description: Image reference to move (required).
          example: docker.io/library/nginx:1.27
          type: string
        to:
          description: Destination registry store to copy into; empty means engines
            pull from `from` directly.
          example: local-cache
          type: string
      required:
      - ref
      type: object
    server.errorResponse:
      properties:
        error:
          type: string
      type: object
...
    server.jobListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/warm.JobSnapshot'
          type: array
          uniqueItems: false
      type: object
...
    warm.JobSnapshot:
      properties:
        created_at:
          type: string
        ended_at:
          type: string
        error:
          type: string
        id:
          type: string
        platforms:
          items:
            type: string
          type: array
          uniqueItems: false
        ref:
          type: string
        started_at:
          type: string
        state:
          $ref: '#/components/schemas/warm.JobState'
        transfers:
          items:
            $ref: '#/components/schemas/warm.TransferSnapshot'
          type: array
          uniqueItems: false
      type: object
    warm.JobState:
      enum:
      - pending
      - running
      - done
      - failed
      - canceled
      type: string
      x-enum-varnames:
      - JobPending
      - JobRunning
      - JobDone
      - JobFailed
      - JobCanceled
    warm.LayerSnapshot:
      properties:
        digest:
          type: string
        done:
          type: integer
        platform:
          type: string
        state:
          description: per-layer progress state
          enum:
          - pending
          - pulling
          - done
          - exists
          - failed
          type: string
        total:
          type: integer
      type: object
    warm.TransferSnapshot:
      properties:
        bytes_done:
          type: integer
        bytes_total:
          type: integer
        error:
          type: string
        from:
          type: string
        kind:
          description: which store kind ran this step
          enum:
          - oci
          - docker
          - containerd
          type: string
        layers:
          items:
            $ref: '#/components/schemas/warm.LayerSnapshot'
          type: array
          uniqueItems: false
        ref:
          type: string
        state:
          description: transfer step state
          enum:
          - pending
          - running
          - done
          - exists
          - failed
          type: string
        store:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: |-
    Move container images between stores (OCI registries and docker/containerd engines) and track per-layer progress.
    A job copies an image `from` an OCI store `to` another, then `distribute` engines pull it.
  title: gantry API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /v1/job:
    get:
      parameters:
      - description: filter by state
        in: query
        name: state
        schema:
          type: string
      - description: filter by ref substring
        in: query
        name: ref
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.jobListResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: List jobs
      tags:
      - jobs
    post:
      description: 'Move an image: copy `from` (oci) into `to` (oci), then have the
        `distribute` engines pull it. Idempotent per identical move.'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.createJobRequest'
        description: job request
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/warm.JobSnapshot'
          description: Accepted
          headers:
            Location:
              description: canonical URL of the created job (/v1/job/{id})
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Request
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Service Unavailable
      security:
      - BearerAuth: []
      summary: Create a job
      tags:
      - jobs
  /v1/job/{id}:
    delete:
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Cancel or evict a job
      tags:
      - jobs
    get:
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/warm.JobSnapshot'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get a job
      tags:
      - jobs
  /v1/job/{id}/progress:
    get:
      description: 'Streams Server-Sent Events: repeated `event: progress` frames
        each carrying a JSON warm.JobSnapshot in `data:`, ending with a terminal `event:
        done` frame. With ?wait=<dur> it instead long-polls and returns a single JSON
        warm.JobSnapshot (no SSE framing) once the job is terminal or the wait elapses.'
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: long-poll duration, e.g. 30s
        in: query
        name: wait
        schema:
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/warm.JobSnapshot'
          description: OK
        "404":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "500":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Stream job progress
      tags:
      - jobs
...